refactor(rust): move tests to dir.rs and fix root namespace handling
- Moves tests from separate test file into dir.rs as a tests module
- Fixes namespace operations to properly handle root namespace:
  - list_namespaces returns empty list for root namespace
  - describe_namespace succeeds for root namespace
  - namespace_exists always succeeds for root namespace
  - create_namespace fails for root (already exists) and non-root (not supported)
  - drop_namespace fails for root (cannot be dropped) and non-root (not supported)
- Updates tests to verify correct root namespace behavior
